# perf: fix: tx/textual/valuerender: use io.WriteString to skip str->byteslice + fix negative sign dropping