… tx/textual/internal/testdata/decimals.json